Widen abilities to u16 and add the hidden-ability slot
Phase 4 element 2. hg-engine's ability ids run past 255 and the donor ROM
already stores them two bytes wide, so every place this engine held an
ability in a byte had to grow before any of those abilities can be brought
across.

The species record is the visible half. u8 abilities[2] became
u16 abilities[3], which takes pl_personal.narc from 44 bytes a record to 48
and shifts everything after the abilities by four. Slot 2 is the hidden
ability, ABILITY_NONE there means the species has none, and in the species
JSON the third entry is optional so five hundred files did not need an
ABILITY_NONE written into them.

The stored ability is the half that breaks saves. Block A of a boxed Pokemon
is full to the byte, ending in a u32 of ribbon flags, so the ability moved to
block B's spare u16 at 0x1A and block A's byte is now dead.
MON_DATA_UNUSED_114 was the only accessor for that spare and has lost its
cases; its enum member stays so no other parameter's value shifts. On a save
made before this, every Pokemon's ability reads as ABILITY_NONE, so start a
new game. docs/oxide/save-layout.md is the running record of this and
everything like it, which is the file Phase 5 asked for.

Everything else that held an ability in a byte follows: BattleMon,
Battler_Ability's return type, DamageCalcParams, RefreshPartyStatusMessage,
the summary screen, the wild-encounter field params, and the daycare and AI
locals. Two deliberate exceptions. trainer_data.h's u8 ability is a slot
selector rather than an id and is unchanged. FrontierPokemon is a packed data
record and is also unchanged, so the one call site that fed it straight to
the setter now copies it into a u16 first, rather than letting the setter
read a byte of friendship along with it.

Hidden abilities arrive as a mechanism and not a policy.
Pokemon_TryGiveHiddenAbility switches a Pokemon to slot 2 and returns FALSE
when the species has none, and a new GiveHiddenAbility script command exposes
it. Nothing calls it yet: which encounters and gifts hand out hidden
abilities is the Phase 5 encounter-design question, and no species has one
until element 3 brings Hardlove's a/0/2/8 table across.

Verified by rebuilding. All 508 personal records agree with the base ROM
field for field once the four-byte shift is applied. A hidden ability put on
Clefairy read back correctly from the built NARC and was then reverted. The
script disassembler still round-trips 574 of 574 with the new opcode in its
table, and the three bulk dry-runs still report nothing to do.

+++ include/applications/pokemon_summary_screen/main.h
@@ -384,7 +384,7 @@ typedef struct PokemonSummaryMonData {
     u16 spAttack;
     u16 spDefense;
     u16 speed;
-    u8 ability;
+    u16 ability; // Platinum Oxide: u16, ability ids run past 255
     u8 nature;
 
     // The stat viewer's alternate readings, in enum PokemonStat order. They sit
